pyproject.toml
[project]
name = "moat-research"
description = "Solo-operator CLI for discovering, scoring, and graduating data-moat opportunities"
requires-python = ">=3.12"
dynamic = ["version"]
dependencies = [
    "claude-agent-sdk>=0.1.77,<0.2.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "jsonschema>=4.21.0",
    "waybackpy>=3.0.6",
    "rich>=13.7.0",
]

[project.scripts]
mr = "mr.cli.main:app"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "anyio>=4.13.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "mr/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["mr"]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonPath = ".venv/bin/python"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --strict-markers --strict-config"
filterwarnings = ["error"]
anyio_mode = "strict"

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
